Split Admin view tests to respective apps
projectwide app had all AdminViews tests. Move them to hwdoc and
keyvalue respectively. Undo the out of app calls that were used before.
In keyvalue's case replace it with minimal instance creation so that
test pass. In hwdoc's case, inherit a class that provides full fixture
creation functionality already
